Added JedisPool support
Inited JedisPool on activation. Changed getJedis() to return a Pool
resource and added a convenience wrapper for returning the Jedis
resource back to the pool. Moved all used of Jedis in this
implemenation to a catch-and-release pattern, ensuring that recursive
calls dont sit and spin on Jedis resoruces while waiting for recursion
to complete. (removeResource)